downloads/SMU-DAL-FSF-FT-06-2019-E-C-master/SMU-DAL-FSF-FT-06-2019-E-C-master/week-05/day-04/mockito-movie/src/main/java/com/company/mockitomovie/MovieService.java
package com.company.mockitomovie;

import java.util.List;
import java.util.stream.Collectors;

public class MovieService {
    private MovieRepository movieRepo;

    public MovieService(MovieRepository movieRepo) {
        this.movieRepo = movieRepo;
    }

    public List<Movie> searchMyMovie(int id, String name) {
        List<Movie> myMovies = movieRepo.getUserMovieList(id);
        return myMovies.stream()
                .filter(movie -> movie.getName().contains(name))
                .collect(Collectors.toList());
    }
}
